Added some new features
Added command line history support. This is based on he GNU history
library. It does not support all of the history functionality that a
shell does; i.e history expansion, but it supports the important stuff.

Added some initial playlist support. Creating a playlist is possible,
but thats essentially it.

Signed-off-by: Alex Waterman <imNotListening@gmail.com>